# src/api/types.go
package api

type LogMessage struct {
	Message     string `json:"message"`
	ProcessName string `json:"process_name"`
}

// NameResponse represents a simple response containing a process name.
type NameResponse struct {
	Name string `json:"name"`
}

// ProjectNameResponse represents a response containing the project name.
type ProjectNameResponse struct {
	ProjectName string `json:"projectName"`
}

// StatusResponse represents a simple response containing a status string.
type StatusResponse struct {
	Status string `json:"status"`
}

// LogsResponse represents a response containing logs lines.
type LogsResponse struct {
	Logs []string `json:"logs"`
}

type ErrorResponse struct {
	Error string `json:"error"`
}

// SendKeysRequest represents the body for sending keystrokes to a process's stdin.
type SendKeysRequest struct {
	Keys string `json:"keys"`
}